[rename]: 결제 페이지를 선물 결제 페이지로 변경 (#110)
