+++ src/components/feature/PaymentDetail/index.module.scss
